updated main page
