+++ src/views/dashboard/components/handleSubmit.js
@@ -1,0 +1,21 @@
+import axios from 'axios'
+
+const handleSubmit = (userObj, setModalIsOpen) => {
+  axios
+    .post(`${process.env.REACT_APP_URL}/register-user`, userObj)
+    .then((res) => {
+      setModalIsOpen(false)
+      window.location.reload()
+    })
+    .catch((err) => {
+      console.log(err)
+      if (err.response.status === 401) {
+        localStorage.clear()
+        window.location.href = '/#/login'
+      } else if (err.response.status === 409) {
+        alert('Polaznik sa ovim emailom već postoji')
+      }
+    })
+}
+
+export default handleSubmit
